Bouncer_API/requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
dj-database-url==0.5.0
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.2.0
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
